Incremental learning transfer
I was looking for a solution of the type proposed since the quality in the identification of the Portuguese language is very poor, being often confused with Spanish, Galician and Catalan. Although in some cases it can be confused, this should not affect the probability of being Portuguese as it is today. eric.x.sun (https://github.com/ericxsun) started this changes before but they were not accepted.